feat: add account balance and token total counts
