+++ src/block.ts
@@ -1,8 +1,7 @@
-import { Address, ethereum } from '@graphprotocol/graph-ts'
-import { Contract } from "./types/Contract/Contract"
-import { Account } from "./types/schema"
+import { ethereum, BigInt } from '@graphprotocol/graph-ts'
 import { getTokenInstance } from "./token"
-import { convertTokenToDecimal } from "./helpers"
+import { updateAccountBalances } from "./accounts"
+
 
 // This is invoked every block, after all event and call handlers
 export function handleBlock(block: ethereum.Block ) : void {
@@ -11,19 +10,11 @@ export function handleBlock(block: ethereum.Block ) : void {
 
 // This is invoked every block that has a call to the SIFI contract, after all event and call handlers
 export function handleBlockWithContractCall(block: ethereum.Block ) : void {
+  let totalHolders = updateAccountBalances()
+
+  // Update total counts
   let token = getTokenInstance()
-
-  token.accounts.forEach((accountId) => {
-    let account = Account.load(accountId)
-    if (account == null) {
-      return
-    }
-
-    let address = Address.fromString(accountId)
-    let contract = Contract.bind(address)
-
-    // Update account balance from contract
-    account.balance = convertTokenToDecimal(contract.balanceOf(address))
-    account.save()
-  })
+  token.totalHolders = totalHolders
+  token.totalTransfers = BigInt.fromI32(token.transfers.length)
+  token.save()
 }
